paritioner v1.0 for conv only

+++ _check_conv_partitioner.py
@@ -9,7 +9,6 @@ from tqdm import tqdm
 from termcolor import colored
 import warnings
 
-
 def set_seed(seed):
     random.seed(seed)
     torch.manual_seed(seed)
@@ -20,19 +19,24 @@ def parse_argument():
     parser = argparse.ArgumentParser()
     parser.add_argument('--num', type=int, default=20, 
                         help='path to model2.onnx file.')
+    parser.add_argument('--use_cuda', action='store_true',
+                        help="Enable CUDA execution")
+    parser.add_argument('--rtol', type=float, default=1e-2, 
+                        help='The relative tolerance parameter.')
+    parser.add_argument('--atol', type=float, default=1e-3, 
+                        help='The absolute tolerance parameter.')
     args = parser.parse_args()
     return args
 
 
 class ConvOnlyNet(nn.Module):
-    def __init__(self, in_channels, num_layers):
+    def __init__(self, in_channels, num_layers, k):
         super().__init__()
         layers = []
         c = in_channels
 
         for i in range(num_layers):
             out_c = random.choice([i for i in range(16, 128)])
-            k = random.choice([1, 3, 5, 7, 9, 11, 13])  # odd kernels
 
             layers.append(nn.Conv2d(c, out_c, kernel_size=k, padding=random.choice([i for i in range(0, k)])))
             # layers.append(nn.ReLU())
@@ -47,6 +51,7 @@ class ConvOnlyNet(nn.Module):
 
 
 if __name__ == '__main__':
+    device = torch.device("cuda" if torch.cuda.is_available() else "mps" if torch.backends.mps.is_available() else "cpu")
     warnings.filterwarnings("ignore", category=FutureWarning)
     # -----------------------------
     # Create output folder
@@ -59,14 +64,17 @@ if __name__ == '__main__':
         in_channels=random.randint(1, 64)
         img_h = random.randint(16, 512)
         img_w = random.randint(16, 512)
+        k = random.choice([1, 3, 5, 7, 9, 11, 13])
 
         # dummy input and model
-        dummy_input = torch.randn(1, in_channels, img_h, img_w)
-        model = ConvOnlyNet(in_channels=in_channels, num_layers=random.randint(1, 5))
+        dummy_input = torch.randn(1, in_channels, img_h, img_w).to(device)
+        model = ConvOnlyNet(in_channels=in_channels, 
+                            num_layers=random.randint(1, 1),
+                            k=k ).to(device)
 
         # test model
         model.eval()
-        out = model(dummy_input)
+        # out = model(dummy_input)
         onnx_path = f"test/model_{in_channels}_in_{img_h}x{img_w}.onnx"
         partitioned_path = f"test/model_{in_channels}_in_{img_h}x{img_w}_partitioned.onnx"
 
@@ -79,12 +87,19 @@ if __name__ == '__main__':
             output_names=["output"],
             verbose=False
         )
+        
+        # partition
+        p = subprocess.run(["python3", "partitioner.py", onnx_path], stdout=subprocess.DEVNULL, stderr=subprocess.PIPE)
+        if p.returncode != 0 and b"The kernel size is too big. Cannot partition." in p.stderr:
+            continue
 
-        subprocess.run(["python3", "partitioner.py", onnx_path], stdout=subprocess.DEVNULL)
-        p = subprocess.run(["python3", "check.py", onnx_path, partitioned_path], stdout=subprocess.DEVNULL)
+        # check
+        cmds = ["python3", "check.py", "--rtol", str(args.rtol), "--atol", str(args.atol), onnx_path, partitioned_path]
+        if args.use_cuda:
+            cmds.extend("--use_cuda")
+        p = subprocess.run(cmds, stdout=subprocess.DEVNULL)
         if p.returncode != 0:
-            print(colored("Conv2d partitioner consistency check failed", "red"))
-            exit(1)
+            raise RuntimeError("Conv2d partitioner consistency check failed")
         try:
             os.remove(onnx_path)
             os.remove(onnx_path+".data")
